ref: fix remaining 'valid-type' typing errors
